+++ tests/Unit/Stubs/SMO.cs
@@ -1646,7 +1646,7 @@ namespace Microsoft.SqlServer.Management.Smo.Agent
     //  New-SqlDscAgentAlert.Tests.ps1
     //  Set-SqlDscAgentAlert.Tests.ps1
     //  Remove-SqlDscAgentAlert.Tests.ps1
-    //  Test-SqlDscAgentAlert.Tests.ps1
+    //  Test-SqlDscIsAgentAlert.Tests.ps1
     //  SqlAgentAlert.Tests.ps1
     public enum AlertType
     {
@@ -1771,7 +1771,7 @@ namespace Microsoft.SqlServer.Management.Smo.Agent
     //  New-SqlDscAgentAlert.Tests.ps1
     //  Set-SqlDscAgentAlert.Tests.ps1
     //  Remove-SqlDscAgentAlert.Tests.ps1
-    //  Test-SqlDscAgentAlert.Tests.ps1
+    //  Test-SqlDscIsAgentAlert.Tests.ps1
     public class Alert
     {
         // Constructor